requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone-client
google-generativeai